Refactor submodule name.
