+++ docs/reference/summarizing.md
@@ -1,0 +1,8 @@
+# Summarizing
+Functions to summarize (*i.e.* create new tables) WI data.
+
+### ::: wiutils.compute_count_summary
+### ::: wiutils.compute_detection
+### ::: wiutils.compute_detection_history
+### ::: wiutils.compute_general_count
+### ::: wiutils.compute_hill_numbers
