+++ tests/darwincore/test_create_dwc_events.py
@@ -1,5 +1,5 @@
 """
-Test cases for the wiutils.transformation.create_dwc_events function.
+Test cases for the wiutils.summarizing.create_dwc_events function.
 """
 import numpy as np
 import pandas as pd
