+++ tests/summarizing/test_compute_detection_history.py
@@ -1,5 +1,5 @@
 """
-Test cases for the wiutils.transformation.compute_detection_history function.
+Test cases for the wiutils.summarizing.compute_detection_history function.
 """
 import numpy as np
 import pandas as pd
