+++ tests/summarizing/test_compute_general_count.py
@@ -1,5 +1,5 @@
 """
-Test cases for the wiutils.transformation.compute_general_count function.
+Test cases for the wiutils.summarizing.compute_general_count function.
 """
 import pandas as pd
 import pytest
